# packages/contracts/generated/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "FetchBody": {
        "properties": {
          "id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "FetchBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "PrepareContextBody": {
        "properties": {
          "max_items": {
            "default": 32,
            "maximum": 128.0,
            "minimum": 1.0,
            "title": "Max Items",
            "type": "integer"
          },
          "max_tokens": {
            "default": 4096,
            "maximum": 32768.0,
            "minimum": 256.0,
            "title": "Max Tokens",
            "type": "integer"
          },
          "purpose": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purpose"
          },
          "query": {
            "maxLength": 4096,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "PrepareContextBody",
        "type": "object"
      },
      "ProposeMemoryBody": {
        "properties": {
          "assertion_text": {
            "minLength": 1,
            "title": "Assertion Text",
            "type": "string"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          },
          "target_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          },
          "target_type": {
            "default": "memory",
            "maxLength": 64,
            "title": "Target Type",
            "type": "string"
          },
          "title": {
            "default": "MCP proposal",
            "maxLength": 512,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "space_id",
          "assertion_text"
        ],
        "title": "ProposeMemoryBody",
        "type": "object"
      },
...
      "RecordInteractionBody": {
        "properties": {
          "client_conversation_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Client Conversation Id",
            "type": "string"
          },
          "client_turn_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Turn Id"
          },
          "completeness": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Completeness",
            "type": "string"
          },
          "content": {
            "maxLength": 65536,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "context_receipt_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Receipt Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "parent_turn_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Turn Id"
          },
          "role": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Role",
            "type": "string"
          },
          "space_id": {
            "format": "uuid",
            "title": "Space Id",
            "type": "string"
          }
        },
        "required": [
          "space_id",
          "client_conversation_id",
          "role",
          "content",
          "completeness"
        ],
        "title": "RecordInteractionBody",
        "type": "object"
      },
...
      "SearchBody": {
        "properties": {
          "query": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "SearchBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Canonical domain API. Adds documents, memory, context, sources, jobs, and ingestion routes.",
    "title": "Memdot Core API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/mcp/fetch": {
      "post": {
        "operationId": "mcp_fetch_api_v1_mcp_fetch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Fetch",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/prepare-context": {
      "post": {
        "operationId": "mcp_prepare_context_api_v1_mcp_prepare_context_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrepareContextBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Prepare Context",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/propose-memory": {
      "post": {
        "operationId": "mcp_propose_memory_api_v1_mcp_propose_memory_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeMemoryBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Propose Memory",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/record-interaction": {
      "post": {
        "operationId": "mcp_record_interaction_api_v1_mcp_record_interaction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordInteractionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Record Interaction",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/search": {
      "post": {
        "operationId": "mcp_search_api_v1_mcp_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mcp Search",
        "tags": [
          "mcp"
        ]
      }
    },
...
  }
}